Support addressing lights by IP address
Previous to this commit there was no way to control bulbs that were on a
different network that homebridge. The only discovery method was through
network broadcasts, which won't work across VLANs.

The node-lifx library supports addressing bulbs by their IP, but this
homebridge plugin did not. This commit adds support by adding the
"lights" configuration option that accepts an array of IP addresses.